requirements.txt
Django==2.1.5
mysqlclient==1.4.1
awscli==1.16.96
botocore==1.12.86
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
diff-match-patch==20181111
django-filter==2.0.0
django-import-export==1.2.0
djangorestframework==3.9.1
docutils==0.14
et-xmlfile==1.0.1
futures==3.2.0
idna==2.8
jdcal==1.4
jmespath==0.9.3
c==0.8.6
MySQL-python==1.2.5
odfpy==1.4.0
openpyxl==2.5.12
pyasn1==0.4.5
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
